# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: InventorySync API
  description: |
    Enterprise-level inventory management API for Shopify stores at startup-friendly prices.
    
    ## Features
    - 🎯 **Custom Fields**: Unlimited custom fields with JSONB storage
    - 🤖 **Workflow Automation**: Event-driven rules engine  
    - 📊 **Advanced Reporting**: Build reports on any field
    - 🚨 **Smart Alerts**: Template-based alerts with analytics
    - 🔗 **Integrations**: Full REST API for third-party systems
    
    ## Authentication
    All endpoints require either Shopify session authentication or API key authentication via `X-API-Key` header.
  version: 1.0.0
  contact:
    name: InventorySync API Support
    email: api-support@inventorysync.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.inventorysync.com/api
    description: Production server
  - url: https://staging-api.inventorysync.com/api  
    description: Staging server
security:
  - ApiKeyAuth: []
  - ShopifySession: []
paths:
  /custom-fields/{shop_domain}:
    get:
      summary: Get custom fields
      description: Retrieve all custom field definitions for a store
      tags:
        - Custom Fields
      parameters:
        - name: shop_domain
          in: path
          required: true
          schema:
            type: string
          example: "test-store.myshopify.com"
        - name: entity_type
          in: query
          schema:
            type: string
            enum: [product, variant, inventory_item, supplier]
          example: "product"
        - name: include_inactive
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Custom fields retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  fields:
                    type: array
                    items:
                      $ref: '#/components/schemas/CustomField'
                  total_count:
                    type: integer
              example:
                fields:
                  - id: 123
                    field_name: "season"
                    display_name: "Product Season"
                    field_type: "select"
                    target_entity: "product"
                    validation_rules:
                      options: ["Spring", "Summer", "Fall", "Winter"]
                    usage_count: 45
                    created_at: "2024-01-15T10:30:00Z"
                total_count: 1

    post:
      summary: Create custom field
      description: Create a new custom field definition
      tags:
        - Custom Fields
      parameters:
        - name: shop_domain
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomFieldCreate'
            example:
              field_name: "material_type"
              display_name: "Material Type"
              field_type: "select"
              target_entity: "product"
              validation_rules:
                options: ["Cotton", "Polyester", "Wool", "Silk"]
              is_required: false
              industry_template: "fashion"
      responses:
        '201':
          description: Custom field created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomField'
...
  /reports/{shop_domain}/fields/{entity_type}:
    get:
      summary: Get available fields
      description: Get all available fields for reporting on an entity type
      tags:
        - Reports
      parameters:
        - name: shop_domain
          in: path
          required: true
          schema:
            type: string
        - name: entity_type
          in: path
          required: true
          schema:
            type: string
            enum: [product, variant, inventory, alert]
      responses:
        '200':
          description: Available fields retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  entity_type:
                    type: string
                  standard_fields:
                    type: array
                    items:
                      $ref: '#/components/schemas/FieldDefinition'
                  custom_fields:
                    type: array
                    items:
                      $ref: '#/components/schemas/FieldDefinition'
                  total_fields:
                    type: integer
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    ShopifySession:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://shopify.dev/oauth/authorize
          tokenUrl: https://shopify.dev/oauth/token
          scopes:
            read_products: Read products
            write_products: Write products
            read_inventory: Read inventory
            write_inventory: Write inventory
  schemas:
    CustomField:
      type: object
      properties:
        id:
          type: integer
          example: 123
        field_name:
          type: string
          example: "season"
        display_name:
          type: string
          example: "Product Season"
        field_type:
          type: string
          enum: [text, number, date, boolean, select, multi_select, url, email]
          example: "select"
        target_entity:
          type: string
          enum: [product, variant, inventory_item, supplier]
          example: "product"
        validation_rules:
          type: object
          example:
            options: ["Spring", "Summer", "Fall", "Winter"]
        is_required:
          type: boolean
          example: false
        is_active:
          type: boolean
          example: true
        usage_count:
          type: integer
          example: 45
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CustomFieldCreate:
      type: object
      required:
        - field_name
        - display_name
        - field_type
        - target_entity
      properties:
        field_name:
          type: string
          pattern: "^[a-z_][a-z0-9_]*$"
          example: "material_type"
        display_name:
          type: string
          example: "Material Type"
        field_type:
          type: string
          enum: [text, number, date, boolean, select, multi_select, url, email]
        target_entity:
          type: string
          enum: [product, variant, inventory_item, supplier]
        validation_rules:
          type: object
        is_required:
          type: boolean
          default: false
        industry_template:
          type: string
          enum: [fashion, food, electronics, B2B]
...
    FieldDefinition:
      type: object
      properties:
        field:
          type: string
        label:
          type: string
        type:
          type: string
        category:
          type: string
          enum: [standard, custom]
        aggregations:
          type: array
          items:
            type: string
...